Test: Split test file MigrationSetExtensionsTests.cs
